social page, create and search posts

+++ im-server/src/main/java/com/zanghongtu/imserver/controller/dto/mine/ProfileDTO.java
@@ -32,7 +32,7 @@ public class ProfileDTO {
 
     private Date birthDate;
 
-    private Double height;
+    private Integer height;
 
     private Double weight;
